+++ sandbox/src/mrc/08-rgradient-marker/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "~/shared/tailwind.css";
 
-const strict = false;
+const strict = true;
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   strict ? (